+++ sdk/resourcemanager/azure-resourcemanager-network/src/main/java/com/azure/resourcemanager/network/implementation/VipSwapsClientImpl.java
@@ -57,7 +57,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
      */
     @Host("{$host}")
     @ServiceInterface(name = "NetworkManagementCli")
-    private interface VipSwapsService {
+    public interface VipSwapsService {
         @Headers({"Content-Type: application/json"})
         @Get(
             "/subscriptions/{subscriptionId}/resourceGroups/{groupName}/providers/Microsoft.Compute/cloudServices"
@@ -140,7 +140,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
         final String singletonResource = "swap";
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(
@@ -193,7 +193,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
         final String singletonResource = "swap";
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
@@ -298,7 +298,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
             parameters.validate();
         }
         final String singletonResource = "swap";
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(
@@ -357,7 +357,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
             parameters.validate();
         }
         final String singletonResource = "swap";
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service
@@ -560,7 +560,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         return FluxUtil
             .withContext(
@@ -611,7 +611,7 @@ public final class VipSwapsClientImpl implements VipSwapsClient {
                     new IllegalArgumentException(
                         "Parameter this.client.getSubscriptionId() is required and cannot be null."));
         }
-        final String apiVersion = "2022-05-01";
+        final String apiVersion = "2022-07-01";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service